remove bash_profile, use profile instead
Initializations run in bash_profile weren't propagating to subsequent
shells. This drove me to put basically everything in bashrc. This is not
ideal, since you should have a place for one-time initializations (e.g.
setting env vars) that's separate from the bells and whistles you want
loaded into interactive shells. The bells and whistles belong in bashrc.

When I installed the Nix package manager, it put some one-time
initialization into bash_profile, but since it never propagated to
subsequent interactive shells, it was useless. Moving this sort of stuff
to ~/.profile seems to fix the problem. My hypothesis is that when we
initially log into Ubuntu or Gnome (I'm still fuzzy as to whether this
is happening at the OS level or the graphical shell level), it runs
~/.profile in a way that propagates to all subsequent shells, whereas
bash_profile is either not run on startup or run in such a way that
propagation is avoided.

Next, start moving stuff that doesn't belong in bashrc into profile.