alacritty light/dark consistency
Let the light/dark state of alacritty.yml drive the THEME environment
variable's default value. Furthermore, use a git filter to ignore
changes to the 'colors' key (light/dark) for the purposes of verison
control. In order to integrate this filter with the repo's local config,
it's necessary to run:

  git config --local include.path ../.gitconfig

Otherwise, the config in .git/config won't be aware of the .gitconfig at
the repo's top-level.