tmux: C-, and zoomed pane navigation
Recognize C-, and make some bindngs to navigate zoomed panes similar to
how iTerm navigates zoomed windows (staying zoomed, but changing which
window is visible). Also, make the O binding repeatable. This would be
more useful with a repeat-time higher than the default (500ms), but I'm
hesitant to change that, as it could get in the way for more common
use-cases.